+++ tests/client.rs
@@ -17,7 +17,7 @@ fn test_create() {
         .with_body(r#"{"_id":"11111111111111111111","name":"rust","count":42,"_createdOn":"2019-09-22T12:24:37.513Z"}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let data = Data {
         name: "rust".into(),
         count: 42,
@@ -41,7 +41,7 @@ fn test_read_all() {
         .with_body(r#"[{"_id":"11111111111111111111","name":"kuy","count":42,"_createdOn":"2019-09-23T12:24:37.513Z"},{"_id":"22222222222222222222","name":"github","count":7,"_createdOn":"2019-09-22T12:24:37.513Z"}]"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.read().all::<Data>();
     assert!(res.is_ok());
 
@@ -65,7 +65,7 @@ fn test_read_all_empty() {
         .with_body("[]")
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("99999999999999999999", &server_url);
+    let client = Client::new("99999999999999999999").with_base_url(&server_url);
     let res = client.read().all::<Data>();
     assert!(res.is_ok());
 
@@ -81,7 +81,7 @@ fn test_read_limit() {
         .with_body(r#"[{"_id":"11111111111111111111","name":"kuy","count":42,"_createdOn":"2019-09-23T12:24:37.513Z"}]"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.read().limit(1).run::<Data>();
     assert!(res.is_ok());
 
@@ -101,7 +101,7 @@ fn test_read_skip() {
         .with_body(r#"[{"_id":"22222222222222222222","name":"github","count":7,"_createdOn":"2019-09-22T12:24:37.513Z"}]"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.read().skip(1).run::<Data>();
     assert!(res.is_ok());
 
@@ -121,7 +121,7 @@ fn test_read_sort() {
         .with_body(r#"[{"_id":"22222222222222222222","name":"github","count":7,"_createdOn":"2019-09-22T12:24:37.513Z"},{"_id":"11111111111111111111","name":"kuy","count":42,"_createdOn":"2019-09-23T12:24:37.513Z"}]"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.read().order_by("count").run::<Data>();
     assert!(res.is_ok());
 
@@ -147,7 +147,7 @@ fn test_read() {
         .with_body(r#"{"_id":"11111111111111111111","name":"kuy","count":42,"_createdOn":"2019-09-22T12:24:37.513Z"}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.read().id::<Data>("11111111111111111111");
     assert!(res.is_ok());
 
@@ -167,7 +167,7 @@ fn test_read_unknown_record_id() {
         .with_body(r#"{"message":"Cannot read property '_id' of null"}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.read().id::<Data>("11111111111111111111");
     assert!(res.is_err());
 
@@ -183,7 +183,7 @@ fn test_update() {
         .with_body(r#"{"message":"Record updated."}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let data = Data {
         name: "cargo".into(),
         count: 42,
@@ -216,7 +216,7 @@ fn test_update_unknown_record_id() {
         .with_body(r#"{"message":"Invalid record Id"}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let data = Data {
         name: "crates".into(),
         count: 42,
@@ -236,7 +236,7 @@ fn test_delete() {
         .with_body(r#"{"message":"Record removed."}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.delete("22222222222222222222");
     assert!(res.is_ok());
 }
@@ -249,7 +249,7 @@ fn test_delete_unknown_record_id() {
         .with_body(r#"{"message":"Invalid record Id"}"#)
         .create();
     let server_url = mockito::server_url();
-    let client = Client::with_base_url("00000000000000000000", &server_url);
+    let client = Client::new("00000000000000000000").with_base_url(&server_url);
     let res = client.delete("44444444444444444444");
     assert!(res.is_err());